+++ client/src/app/core/core-services/projector-data.service.ts
@@ -94,7 +94,7 @@ export class ProjectorDataService {
             );
         } catch (e) {
             if (!(e instanceof OfflineError)) {
-                console.log(e);
+                console.error(e);
             }
         }
     }
