+++ client/src/app/site/users/components/user-import/user-import-list.component.ts
@@ -54,7 +54,6 @@ export class UserImportListComponent extends BaseImportListComponentDirective<Us
                 prop: `newEntry.${property}`,
                 type: this.guessType(property as keyof User)
             };
-            console.log('singleColumnDef ', singleColumnDef);
 
             return singleColumnDef;
         });
