+++ server/openslides/global_settings.py
@@ -23,7 +23,7 @@ INSTALLED_APPS = [
     "openslides.mediafiles",
 ]
 
-INSTALLED_PLUGINS = collect_plugins()  # Adds all automaticly collected plugins
+INSTALLED_PLUGINS = collect_plugins()  # Adds all automatically collected plugins
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
@@ -33,7 +33,7 @@ MIDDLEWARE = [
     "django.middleware.csrf.CsrfViewMiddleware",
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
-    "openslides.utils.autoupdate.AutoupdateBundleMiddleware",
+    "openslides.utils.autoupdate_bundle.AutoupdateBundleMiddleware",
 ]
 
 ROOT_URLCONF = "openslides.urls"
